♻️ Moved orchestrator subclassed to new module
The orchestrator subclasses implemented at the moment are implementing
final schemes, i.e., ones that are not made to be reused in larger QEC
codes. To avoid the confusion with templates, they are now moved to
another submodule.
