+++ python_prototype/src/tqec/constructions/corner.py
